LF-1862 pci: pcie_imx: Fix potential integer overflow
Fix Coverity issue 6589702. In get_bus_address, potentially
overflowing expression d << 8 with type pci_dev_t (32 bits, signed)
is evaluated using 32-bit arithmetic, and then used in a context
that expects an expression of type u64.

(cherry picked from commit 9f1f55bf04fc53be768b73a7097c6fbd64288620)

+++ drivers/pci/pcie_imx.c
@@ -607,7 +607,7 @@ static void __iomem *get_bus_address(struct imx_pcie_priv *priv,
 			imx_pcie_atu_outbound_set(priv, PCIE_ATU_REGION_INDEX1,
 					 PCIE_ATU_TYPE_CFG0,
 					 (ulong)priv->cfg_base,
-					 d << 8,
+					 (u64)d << 8,
 					 priv->cfg_size >> 1);
 			va_address = priv->cfg_base;
 		} else {
@@ -615,7 +615,7 @@ static void __iomem *get_bus_address(struct imx_pcie_priv *priv,
 			imx_pcie_atu_outbound_set(priv, PCIE_ATU_REGION_INDEX1,
 					 PCIE_ATU_TYPE_CFG1,
 					 (ulong)priv->cfg1_base,
-					 d << 8,
+					 (u64)d << 8,
 					 priv->cfg_size >> 1);
 			va_address = priv->cfg1_base;
 		}
